Fixed renderer for non dx11 cards
Replaced SDK version with NULL in the method D3D11CreateDeviceAndSwapChain(). This allows the hardware to fall back to previous versions of the D3D sdk.

Also changed to compile vs and ps version 4.0, for support with dx10 cards (could drop it lower for 9 support? if needed).

Added asserts if anything fails on viscraft creation.